support v2 mesh & material
