+++ io_ogre/ogre/material.py
@@ -10,6 +10,7 @@ import tempfile
 import shutil
 import logging
 from itertools import chain
+from .materialv2 import OgreMaterialGeneratorV2, OgreMaterialGeneratorBase
 
 def dot_materials(materials, path=None, separate_files=True):
     """
@@ -36,7 +37,10 @@ def dot_materials(materials, path=None, separate_files=True):
                 if mat is None:
                     continue
                 Report.materials.append( material_name(mat) )
-                generator = OgreMaterialGenerator(mat)
+                if config.get('MESH_V2'):
+                    generator = OgreMaterialGeneratorV2(mat)
+                else:
+                    generator = OgreMaterialGenerator(mat)
                 if kwargs.get('copy_programs', config.get('COPY_SHADER_PROGRAMS')):
                     generator.copy_programs(path)
                 if kwargs.get('touch_textures', config.get('TOUCH_TEXTURES')):
@@ -56,7 +60,10 @@ def dot_material(mat, path, **kwargs):
       * touch_textures - bool. Copy the images along to the material files.
     """
     prefix = kwargs.get('prefix', '')
-    generator = OgreMaterialGenerator(mat, prefix=prefix)
+    if config.get('MESH_V2'):
+        generator = OgreMaterialGeneratorV2(mat)
+    else:
+        generator = OgreMaterialGenerator(mat, prefix=prefix)
     if kwargs.get('copy_programs', config.get('COPY_SHADER_PROGRAMS')):
         generator.copy_programs(path)
     if kwargs.get('touch_textures', config.get('TOUCH_TEXTURES')):
@@ -67,20 +74,13 @@ def dot_material(mat, path, **kwargs):
 
     return generator.material_name
 
-class OgreMaterialGenerator(object):
+class OgreMaterialGenerator(OgreMaterialGeneratorBase):
     def __init__(self, material, prefix=''):
-        self.material = material
+        super(OgreMaterialGenerator, self).__init__(material, prefix='')
         self.w = util.IndentedWriter()
-        self.passes = []
-        self.material_name = material_name(self.material,prefix=prefix)
-
-        if material.node_tree:
-            nodes = shader.get_subnodes( self.material.node_tree, type='MATERIAL_EXT' )
-            for node in nodes:
-                if node.material:
-                    self.passes.append( node.material )
 
     def generate(self):
+
         self.w.line('// %s generated by blender2ogre %s' % (self.material.name, datetime.now())).nl()
         self.generate_header()
         with self.w.iword('material').word(self.material_name).embed():
@@ -273,53 +273,6 @@ class OgreMaterialGenerator(object):
                     idx = find_uv_layer_index( uv_layer )
                     self.w.iword('tex_coord_set').integer(idx)
 
-    def copy_textures(self, target_path):
-        slots = get_image_textures(self.material) + list(chain([get_image_textures(mat) for mat in self.passes]))
-        for slot in slots:
-            self.copy_texture(slot, target_path)
-
-    def copy_texture(self, slot, target_path):
-        if not slot:
-            return
-
-        origin_filepath = util.texture_image_path(slot.texture)
-        if origin_filepath == None:
-            return
-
-        tmp_filepath = None
-        updated_image = False
-        if origin_filepath == '.':
-            # a is a packed png
-            origin_filepath = slot.texture.image.filepath
-            _, ext = splitext(origin_filepath)
-            tmp_filepath = tempfile.mkstemp(suffix=ext)
-            slot.texture.image.filepath = tmp_filepath 
-            slot.texture.image.save()
-            slot.texture.image.filepath = origin_filepath
-            updated_image = True
-
-        _, target_file_ext = split(origin_filepath)
-        target_file, ext = splitext(target_file_ext)
-
-        if not tmp_filepath:
-            _, tmp_filepath = tempfile.mkstemp(suffix=ext)
-
-        target_file_ext = self.change_ext(target_file_ext, slot.texture.image)
-        target_filepath = join(target_path, target_file_ext)
-        if not os.path.isfile(target_filepath):
-            # or os.stat(target_filepath).st_mtime < os.stat( origin_filepath ).st_mtime:
-            updated_image = True
-            shutil.copyfile(origin_filepath, tmp_filepath)
-        else:
-            logging.info("skip copy (%s). texture is already up to date.", origin_filepath)
-
-        if updated_image:
-            if is_image_postprocessed(slot.texture.image):
-                logging.info("magick (%s) -> (%s)", tmp_filepath, origin_filepath)
-                util.image_magick(slot.texture, tmp_filepath, target_filepath)
-            else:
-                shutil.copyfile(tmp_filepath, target_filepath)
-                logging.info("copy (%s)", origin_filepath)
 
     def get_active_programs(self):
         r = []
@@ -336,13 +289,6 @@ class OgreMaterialGenerator(object):
             else:
                 logging.warn('uses program %s which has no source' % (prog.name))
 
-    def change_ext( self, name, image ):
-        name_no_ext, _ = splitext(name)
-        if image.convert_format != 'NONE':
-            name = name_no_ext + "." + image.convert_format
-        if config.get('FORCE_IMAGE_FORMAT') != 'NONE':
-            name = name_no_ext + "." + config.get('FORCE_IMAGE_FORMAT')
-        return name
 
 # Make default material for missing materials:
 # * Red flags for users so they can quickly see what they forgot to assign a material to.
@@ -397,27 +343,6 @@ def load_user_materials():
         for prog in progs:
             logging.info('Ogre shader program', prog.name)
 
-
-def material_name( mat, clean = False, prefix='' ):
-    """
-    returns the material name.
-
-    materials from a library might be exported several times for multiple objects.
-    there is no need to have those textures + material scripts several times. thus
-    library materials are prefixed with the material filename. (e.g. test.blend + diffuse
-    should result in "test_diffuse". special chars are converted to underscore.
-
-    clean: deprecated. do not use!
-    """
-    if type(mat) is str:
-        return prefix + clean_object_name(mat)
-    name = clean_object_name(mat.name)
-    if mat.library:
-        _, filename = split(mat.library.filepath)
-        prefix, _ = splitext(filename)
-        return prefix + "_" + name
-    else:
-        return prefix + name
 
 def get_shader_program( name ):
     if name in OgreProgram.PROGRAMS:
@@ -647,13 +572,6 @@ IMAGE_FORMATS =  [
     ('tga', 'tga', 'targa format'),
     ('dds', 'dds', 'nvidia dds format'),
 ]
-
-def is_image_postprocessed( image ):
-    if config.get('FORCE_IMAGE_FORMAT') != 'NONE' or image.use_resize_half or image.use_resize_absolute or image.use_color_quantize or image.use_convert_format:
-        return True
-    else:
-        return False
-
 
 def update_parent_material_path( path ):
     ''' updates RNA '''
